add route

+++ StaticFileServer/Controllers/ImageFilesController.cs
@@ -1,12 +1,7 @@
-﻿using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
-using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Net.Http;
 using System.Threading.Tasks;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
@@ -18,23 +13,23 @@ namespace StaticFileServer.Controllers
     public class ImageFilesController : ControllerBase
     {
         private IConfiguration _configuration;
-        private string orgPath, filePath;
+        private string orgPath;
         public ImageFilesController(IConfiguration configuration)
         {
             _configuration = configuration;
 
             orgPath = Path.Combine(_configuration["StaticFilePath"]);
-            filePath = Path.Combine(orgPath, "images");
             if (!Directory.Exists(orgPath)) Directory.CreateDirectory(orgPath);
-            if (!Directory.Exists(filePath)) Directory.CreateDirectory(filePath);
         }
 
         // GET api/<ImageFilesController>/5
-        [HttpGet("{imageName}")]
-        public IActionResult Get(string imageName)
+        [HttpGet("{groupName}/{imageName}")]
+        public IActionResult Get(string groupName,string imageName)
         {
-            var imagePath = Path.Combine(filePath, imageName);
+            var imagePath = Path.Combine(orgPath, groupName);
+            if (!Directory.Exists(imagePath)) Directory.CreateDirectory(imagePath);
 
+            imagePath = Path.Combine(imagePath, imageName);
             if (System.IO.File.Exists(imagePath))
             {
                 return (PhysicalFile(imagePath, "image/jpeg"));
@@ -44,16 +39,17 @@ namespace StaticFileServer.Controllers
         }
 
         // PUT api/<ImageFilesController>/5
-        [HttpPut("{imageName}")]
+        [HttpPut("{groupName}/{imageName}")]
         //[Consumes("multipart/form-data")]
-        async public Task<IActionResult> Put(string imageName, [FromForm] IFormFile formFile)
+        async public Task<IActionResult> Put(
+            string groupName,string imageName, [FromForm] IFormFile formFile)
         {
-            var imagePath = Path.Combine(filePath, imageName);
+            var imagePath = Path.Combine(orgPath,groupName);
+            if (!Directory.Exists(imagePath)) Directory.CreateDirectory(imagePath);
 
-            if (System.IO.File.Exists(imagePath))
-            {
-                System.IO.File.Delete(imagePath);
-            }
+            imagePath = Path.Combine(imagePath, imageName);
+
+            if (System.IO.File.Exists(imagePath)) System.IO.File.Delete(imagePath);
 
             using (var stream = System.IO.File.Create(imagePath))
             {
@@ -64,16 +60,17 @@ namespace StaticFileServer.Controllers
         }
 
         // DELETE api/<ImageFilesController>/5
-        [HttpDelete("{imageName}")]
-        public IActionResult Delete(string imageName)
+        [HttpDelete("{groupName}/{imageName}")]
+        public IActionResult Delete(string groupName, string imageName)
         {
-            var imagePath = Path.Combine(filePath, imageName);
+            var imagePath = Path.Combine(orgPath, groupName);
+            if (!Directory.Exists(imagePath)) return NotFound();
 
-            if (System.IO.File.Exists(imagePath))
-            {
-                System.IO.File.Delete(imagePath);
-            }
+            imagePath = Path.Combine(imagePath, imageName);
 
+            if (!System.IO.File.Exists(imagePath)) return NotFound();
+
+            System.IO.File.Delete(imagePath);
             return Ok();
         }
     }
